feat: add participants components folder
