+++ src/components/index.ts
@@ -1,11 +1,5 @@
 export * from "./sidebar"
-export * from "./rounds-list"
-export * from "./participants-list"
-export * from "./participant-image"
-export * from "./participant-item"
-export * from "./participant-name"
-export * from "./participants-title"
-export * from "./participants-sub-title"
+export * from "./participants"
 export * from "./sidebar"
 export * from "./card"
 export * from "./button"
