--- src/components/participant-item/participant-item.tsx
@@ -1,5 +1,0 @@
-import { PropsWithChildren } from "@/types/react";
-
-export function ParticipantItem({ children }: PropsWithChildren) {
-  return <li className="flex flex-col justify-center ml-2">{children}</li>;
-}
