--- src/components/participant-name/participant-name.tsx
@@ -1,9 +1,0 @@
-import { PropsWithChildren } from "@/types/react";
-
-export function ParticipantName({ children }: PropsWithChildren) {
-  return (
-    <span className="text-neutral-900 text-xs font-semibold leading-normal block mt-[12px]">
-      {children}
-    </span>
-  );
-}
