--- src/components/participants-list/index.ts
@@ -1,1 +1,0 @@
-export * from "./participants-list";
