--- src/components/participants-list/participants-list.tsx
@@ -1,9 +1,0 @@
-import { PropsWithChildren } from "@/types/react";
-
-export function ParticipantsList({ children }: PropsWithChildren) {
-  return (
-      <ul className="inline-flex mt-[12px]">
-        {children}
-      </ul>
-  )
-}
