+++ src/components/participants/index.ts
@@ -1,0 +1,4 @@
+export * from "./participants-list"
+export * from "./rounds-list"
+export * from "./sub-title"
+export * from "./title"
