+++ src/components/participants/participants-list/participant-image/index.ts
