+++ src/components/participants/participants-list/participant-item/index.ts
