+++ src/components/participants/participants-list/participant-item/participant-item.tsx
@@ -1,0 +1,5 @@
+import { PropsWithChildren } from "react"
+
+export function ParticipantItem({ children }: PropsWithChildren) {
+  return <li className="flex flex-col justify-center ml-2">{children}</li>
+}
