+++ src/components/participants/participants-list/participants-list.tsx
@@ -1,0 +1,5 @@
+import { PropsWithChildren } from "react"
+
+export function ParticipantsList({ children }: PropsWithChildren) {
+  return <ul className="inline-flex mt-[12px]">{children}</ul>
+}
