+++ src/components/participants/rounds-list/index.ts
@@ -1,0 +1,2 @@
+export * from "./rounds-list"
+export * from "./name"
