+++ src/components/participants/rounds-list/name/index.ts
@@ -1,0 +1,1 @@
+export * from "./name"
