+++ src/components/participants/sub-title/index.ts
@@ -1,0 +1,1 @@
+export * from "./sub-title"
