+++ src/components/participants/title/index.ts
@@ -1,0 +1,1 @@
+export * from "./title"
