--- src/components/rounds-list/index.ts
@@ -1,1 +1,0 @@
-export * from "./rounds-list";
